Upload source code finally!
This is actually version 1.1... it's a bit of a funny story, actually. I tried to commit this code to GitHub back in March when I made the program, but I screwed up, and lost everything. Like unrecovarably lost everything. Not on my hard drive, not online, not anywhere. All I had left was a single .exe file. So I decompiled my own .exe, got the source code back, and recreated the program. While doing so, I also made a few tweaks: thus, version 1.1.